+++ src/pages/components/Logo.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Logo = () => {
+  return <img alt="Logo" src="images/mifiel_logo.webp" className="w-32" />;
+};
+
+export default Logo;