+++ src/redux/sagas.js
@@ -1,0 +1,7 @@
+import { all } from "redux-saga/effects";
+import user from "./user/saga";
+import documents from "./documents/saga";
+
+export default function* rootSaga() {
+  yield all([user()]);
+}