+++ src/redux/store.js
@@ -1,9 +1,20 @@
-import { createStore } from "redux";
-import { devToolsEnhancer } from "redux-devtools-extension";
+import { createStore, applyMiddleware } from "redux";
+import { composeWithDevTools } from "redux-devtools-extension";
+import createSagaMiddleware, { runSaga } from "redux-saga";
 import reducers from "./reducers";
+import sagas from "./sagas";
 
-export function configureStore(initialState) {
-  const store = createStore(reducers, initialState, devToolsEnhancer());
+const sagaMiddleware = createSagaMiddleware();
+
+const middlewares = [sagaMiddleware];
+
+export function configureStore() {
+  const store = createStore(
+    reducers,
+    composeWithDevTools(applyMiddleware(...middlewares))
+  );
+
+  sagaMiddleware.run(sagas);
 
   if (module.hot) {
     module.hot.accept("./reducers", () => {